Time: 0 ms (100%), Space: 7.7 MB (98.74%) - LeetHub

+++ 1817-calculate-money-in-leetcode-bank/1817-calculate-money-in-leetcode-bank.cpp
@@ -31,7 +31,7 @@ public:
 
         int rem=n%7;
         n=n/7;
-        long long int num=28*n+(n*(n-1)*7)/2;
+        int num=28*n+(n*(n-1)*7)/2;
         int day=n+1;
         num+=(rem*(rem+1))/2+(day-1)*rem;
         return num;
